+++ lib/main.dart
@@ -14,6 +14,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart' as dotenv;
 import 'package:audio_service/audio_service.dart';
 import 'package:djsports/data/services/djaudio_handler.dart'; // Legg til denne importen
 import 'package:djsports/core/theme/app_theme.dart';
+import 'package:djsports/data/provider/theme_color_provider.dart';
 import 'package:toastification/toastification.dart';
 
 final rootScaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
@@ -79,11 +80,12 @@ class DJSportsApp extends ConsumerWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final themeColor = ref.watch(themeColorProvider);
     return ToastificationWrapper(
       child: MaterialApp(
       title: 'djSports',
       debugShowCheckedModeBanner: false,
-      theme: AppTheme.lightTheme,
+      theme: AppTheme.themeFor(themeColor),
       // reload" button in a Flutter-supported IDE, or press "r" if you used
       // the command line to start the app).
       //